Project structure improvements
